move stories to stories/Input
