+++ packages/react-components/react-input/src/stories/Input/InputControlled.stories.tsx
@@ -2,8 +2,8 @@ import * as React from 'react';
 import { Label } from '@fluentui/react-label';
 import { useId } from '@fluentui/react-utilities';
 import { makeStyles, shorthands } from '@griffel/react';
-import { Input } from '../index';
-import type { InputProps } from '../index';
+import { Input } from '../../index';
+import type { InputProps } from '../../index';
 
 const useStyles = makeStyles({
   root: {
