+++ packages/react-components/react-input/src/stories/Input/InputDefault.stories.tsx
@@ -3,8 +3,8 @@ import { ArgTypes } from '@storybook/api';
 import { Label } from '@fluentui/react-label';
 import { useId } from '@fluentui/react-utilities';
 import { makeStyles, shorthands } from '@griffel/react';
-import { Input } from '../index';
-import type { InputProps } from '../index';
+import { Input } from '../../index';
+import type { InputProps } from '../../index';
 
 const useStyles = makeStyles({
   root: {
