+++ packages/react-components/react-input/src/stories/Input/InputPlaceholder.stories.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { Input } from '../index';
+import { Input } from '../../index';
 import { Label } from '@fluentui/react-label';
 import { useId } from '@fluentui/react-utilities';
 import { makeStyles } from '@griffel/react';
