+++ packages/react-components/react-input/src/stories/Input/InputSize.stories.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { Label } from '@fluentui/react-label';
 import { useId } from '@fluentui/react-utilities';
 import { makeStyles, shorthands } from '@griffel/react';
-import { Input } from '../index';
+import { Input } from '../../index';
 
 const useStyles = makeStyles({
   root: {
